+++ TimberWinR/Inputs/LogsFileDatabase.cs
@@ -5,10 +5,8 @@ using System.Linq;
 using System.Text;
 using Newtonsoft.Json;
 using NLog;
+using TimberWinR.Parser;
 
-using LogQuery = Interop.MSUtil.LogQueryClassClass;
-using TextLineInputFormat = Interop.MSUtil.COMTextLineInputContextClass;
-using LogRecordSet = Interop.MSUtil.ILogRecordset;
 
 namespace TimberWinR.Inputs
 {
@@ -19,7 +17,7 @@ namespace TimberWinR.Inputs
     {
         private static readonly object _locker = new object();
         private List<LogsFileDatabaseEntry> Entries { get; set; }
-        private string DatabaseDirectory { get; set; }      
+        private string DatabaseDirectory { get; set; }
         public string DatabaseFileName
         {
             get { return Path.Combine(DatabaseDirectory, ".timberwinrdb"); }
@@ -37,6 +35,14 @@ namespace TimberWinR.Inputs
             }
         }
 
+        private LogsFileDatabaseEntry FindFile(string logName)
+        {
+            lock (_locker)
+            {
+                var existingEntry = (from e in Entries where e.FileName == logName select e).FirstOrDefault();
+                return existingEntry;
+            }
+        }
         private bool ExistingFileTest(string logName)
         {
             var existingEntry = (from e in Entries where e.FileName == logName select e).FirstOrDefault();
@@ -56,37 +62,51 @@ namespace TimberWinR.Inputs
             }
         }
 
-        private LogsFileDatabaseEntry AddFileEntry(string logName, TextLineInputFormat fmt)
+        private LogsFileDatabaseEntry AddFileEntry(string logName)
         {
             var de = new LogsFileDatabaseEntry();
             lock (_locker)
             {
-                var lq = new LogQuery();
+                de.NewFile = true;
                 var fi = new FileInfo(logName);
                 de.FileName = logName;
                 de.Size = fi.Length;
                 de.SampleTime = DateTime.UtcNow;
-                de.CreationTime = fi.CreationTimeUtc;
-                if (fi.Exists)
-                {
-                    var qcount = string.Format("SELECT max(Index) as MaxRecordNumber FROM {0}", logName);
-                    var rcount = lq.Execute(qcount, fmt);
-                    var qr = rcount.getRecord();
-                    var lrn = (Int64)qr.getValueEx("MaxRecordNumber");
-                    de.MaxRecords = lrn;
-                }
+                de.CreationTimeUtc = fi.CreationTimeUtc;
                 Entries.Add(de);
                 WriteDatabaseFileNoLock();
             }
             return de;
         }
 
-        public static LogsFileDatabaseEntry AddLogFile(string logName, TextLineInputFormat fmt)
+        public static LogsFileDatabaseEntry LookupLogFile(string logName)
         {
-            Instance.RemoveFileEntry(logName); // Remove if already exists, otherwise ignores.
-            return Instance.AddFileEntry(logName, fmt);
+            LogsFileDatabaseEntry dbe = Instance.FindFile(logName);
+            if (dbe == null)
+                dbe = Instance.AddFileEntry(logName);
+            else
+                dbe.NewFile = false;
+    
+            return dbe;
         }
 
+        public static void Update(LogsFileDatabaseEntry dbe)
+        {
+            Instance.UpdateEntry(dbe);
+        }
+
+        private void UpdateEntry(LogsFileDatabaseEntry dbe)
+        {
+            lock(_locker)
+            {
+                var fi = new FileInfo(dbe.FileName);
+                dbe.CreationTimeUtc = fi.CreationTimeUtc;
+                dbe.SampleTime = DateTime.UtcNow;
+                dbe.Size = fi.Length;
+                WriteDatabaseFileNoLock();
+            }
+            
+        }
         public static LogsFileDatabase Instance
         {
             get
@@ -122,6 +142,17 @@ namespace TimberWinR.Inputs
             {
                 LogManager.GetCurrentClassLogger()
                     .Error("Error reading database '{0}': {1}", DatabaseFileName, ex.ToString());
+                try
+                {
+                    if (File.Exists(DatabaseFileName))
+                        File.Delete(DatabaseFileName);
+                    LogManager.GetCurrentClassLogger().Info("Creating New Database '{0}'", DatabaseFileName);
+                    WriteDatabaseLock();
+                }
+                catch (Exception ex2)
+                {
+                    LogManager.GetCurrentClassLogger().Info("Error Creating New Database '{0}': {1}", DatabaseFileName, ex2.ToString());
+                }              
             }
         }
         private void WriteDatabaseFileNoLock()
@@ -155,7 +186,7 @@ namespace TimberWinR.Inputs
         }
 
         private LogsFileDatabase(string databaseDirectory)
-        {         
+        {
             DatabaseDirectory = databaseDirectory;
             Entries = new List<LogsFileDatabaseEntry>();
         }
@@ -164,11 +195,13 @@ namespace TimberWinR.Inputs
 
     public class LogsFileDatabaseEntry
     {
+        [JsonIgnore]
+        public bool NewFile { get; set; }
         public string FileName { get; set; }
         public Int64 MaxRecords { get; set; }
-        public DateTime CreationTime { get; set; }
+        public DateTime CreationTimeUtc { get; set; }
         public DateTime SampleTime { get; set; }
-        public long Size { get; set; }
+        public long Size { get; set; }       
     }
 
 }